Update to latest code changes

+++ src/mongo/db/catalog/database.cpp
@@ -50,12 +50,12 @@
 #include "mongo/db/concurrency/write_conflict_exception.h"
 #include "mongo/db/dbhelpers.h"
 #include "mongo/db/index/index_access_method.h"
-#include "mongo/db/instance.h"
 #include "mongo/db/introspect.h"
 #include "mongo/db/op_observer.h"
 #include "mongo/db/query/collation/collator_factory_interface.h"
 #include "mongo/db/repl/oplog.h"
 #include "mongo/db/repl/replication_coordinator_global.h"
+#include "mongo/db/server_options.h"
 #include "mongo/db/server_parameters.h"
 #include "mongo/db/service_context.h"
 #include "mongo/db/service_context_d.h"
@@ -145,7 +145,7 @@ void Database::close(OperationContext* txn) {
     repl::oplogCheckCloseDatabase(txn, this);
 
     if (BackgroundOperation::inProgForDb(_name)) {
-        log() << "warning: bg op in prog during close db? " << _name << endl;
+        log() << "warning: bg op in prog during close db? " << _name;
     }
 }
 
@@ -203,11 +203,12 @@ Database::Database(OperationContext* txn, StringData name, DatabaseCatalogEntry*
       _dbEntry(dbEntry),
       _profileName(_name + ".system.profile"),
       _indexesName(_name + ".system.indexes"),
-      _viewsName(_name + ".system.views"),
-      _views(txn, this) {
+      _viewsName(_name + "." + DurableViewCatalog::viewsCollectionName().toString()),
+      _durableViews(DurableViewCatalogImpl(this)),
+      _views(&_durableViews) {
     Status status = validateDBName(_name);
     if (!status.isOK()) {
-        warning() << "tried to open invalid db: " << _name << endl;
+        warning() << "tried to open invalid db: " << _name;
         uasserted(10028, status.toString());
     }
 
@@ -219,35 +220,17 @@ Database::Database(OperationContext* txn, StringData name, DatabaseCatalogEntry*
         const string ns = *it;
         _collections[ns] = _getOrCreateCollectionInstance(txn, ns);
     }
-}
-
-
-/*static*/
-string Database::duplicateUncasedName(const string& name, set<string>* duplicates) {
-    if (duplicates) {
-        duplicates->clear();
+    // At construction time of the viewCatalog, the _collections map wasn't initialized yet, so no
+    // system.views collection would be found. Now we're sufficiently initialized, signal a version
+    // change. Also force a reload, so if there are problems with the catalog contents as might be
+    // caused by incorrect mongod versions or similar, they are found right away.
+    _views.invalidate();
+    Status reloadStatus = _views.reloadIfNeeded(txn);
+    if (!reloadStatus.isOK()) {
+        warning() << "Unable to parse views: " << redact(reloadStatus)
+                  << "; remove any invalid views from the " << _viewsName
+                  << " collection to restore server functionality." << startupWarningsLog;
     }
-
-    set<string> allShortNames;
-    dbHolder().getAllShortNames(allShortNames);
-
-    for (const auto& dbname : allShortNames) {
-        if (strcasecmp(dbname.c_str(), name.c_str()))
-            continue;
-
-        if (strcmp(dbname.c_str(), name.c_str()) == 0)
-            continue;
-
-        if (duplicates) {
-            duplicates->insert(dbname);
-        } else {
-            return dbname;
-        }
-    }
-    if (duplicates) {
-        return duplicates->empty() ? "" : *duplicates->begin();
-    }
-    return "";
 }
 
 void Database::clearTmpCollections(OperationContext* txn) {
@@ -269,7 +252,7 @@ void Database::clearTmpCollections(OperationContext* txn) {
             WriteUnitOfWork wunit(txn);
             Status status = dropCollection(txn, ns);
             if (!status.isOK()) {
-                warning() << "could not drop temp collection '" << ns << "': " << status;
+                warning() << "could not drop temp collection '" << ns << "': " << redact(status);
                 continue;
             }
 
@@ -310,7 +293,8 @@ void Database::getStats(OperationContext* opCtx, BSONObjBuilder* output, double 
     list<string> collections;
     _dbEntry->getCollectionNamespaces(&collections);
 
-    long long ncollections = 0;
+    long long nCollections = 0;
+    long long nViews = 0;
     long long objects = 0;
     long long size = 0;
     long long storageSize = 0;
@@ -325,7 +309,7 @@ void Database::getStats(OperationContext* opCtx, BSONObjBuilder* output, double 
         if (!collection)
             continue;
 
-        ncollections += 1;
+        nCollections += 1;
         objects += collection->numRecords(opCtx);
         size += collection->dataSize(opCtx);
 
@@ -337,7 +321,10 @@ void Database::getStats(OperationContext* opCtx, BSONObjBuilder* output, double 
         indexSize += collection->getIndexSize(opCtx);
     }
 
-    output->appendNumber("collections", ncollections);
+    getViewCatalog()->iterate(opCtx, [&](const ViewDefinition& view) { nViews += 1; });
+
+    output->appendNumber("collections", nCollections);
+    output->appendNumber("views", nViews);
     output->appendNumber("objects", objects);
     output->append("avgObjSize", objects == 0 ? 0 : double(size) / double(objects));
     output->appendNumber("dataSize", size / scale);
@@ -349,15 +336,15 @@ void Database::getStats(OperationContext* opCtx, BSONObjBuilder* output, double 
     _dbEntry->appendExtraStats(opCtx, output, scale);
 }
 
+Status Database::dropView(OperationContext* txn, StringData fullns) {
+    Status status = _views.dropView(txn, NamespaceString(fullns));
+    Top::get(txn->getClient()->getServiceContext()).collectionDropped(fullns);
+    return status;
+}
+
 Status Database::dropCollection(OperationContext* txn, StringData fullns) {
-    invariant(txn->lockState()->isDbLockedForMode(name(), MODE_X));
-
-    LOG(1) << "dropCollection: " << fullns << endl;
-    massertNamespaceNotIndex(fullns, "dropCollection");
-
-    Collection* collection = getCollection(fullns);
-    if (!collection) {
-        // collection doesn't exist
+    if (!getCollection(fullns)) {
+        // Collection doesn't exist so don't bother validating if it can be dropped.
         return Status::OK();
     }
 
@@ -370,32 +357,49 @@ Status Database::dropCollection(OperationContext* txn, StringData fullns) {
                 if (_profile != 0)
                     return Status(ErrorCodes::IllegalOperation,
                                   "turn off profiling before dropping system.profile collection");
-            } else {
-                return Status(ErrorCodes::IllegalOperation, "can't drop system ns");
+            } else if (!nss.isSystemDotViews()) {
+                return Status(ErrorCodes::IllegalOperation,
+                              str::stream() << "can't drop system collection " << fullns);
             }
         }
     }
 
+    return dropCollectionEvenIfSystem(txn, nss);
+}
+
+Status Database::dropCollectionEvenIfSystem(OperationContext* txn, const NamespaceString& fullns) {
+    invariant(txn->lockState()->isDbLockedForMode(name(), MODE_X));
+
+    LOG(1) << "dropCollection: " << fullns;
+
+    Collection* collection = getCollection(fullns);
+    if (!collection) {
+        return Status::OK();  // Post condition already met.
+    }
+
+    massertNamespaceNotIndex(fullns.toString(), "dropCollection");
+
     BackgroundOperation::assertNoBgOpInProgForNs(fullns);
 
-    audit::logDropCollection(&cc(), fullns);
+    audit::logDropCollection(&cc(), fullns.toString());
 
     Status s = collection->getIndexCatalog()->dropAllIndexes(txn, true);
     if (!s.isOK()) {
         warning() << "could not drop collection, trying to drop indexes" << fullns << " because of "
-                  << s.toString();
+                  << redact(s.toString());
         return s;
     }
 
     verify(collection->_details->getTotalIndexCount(txn) == 0);
-    LOG(1) << "\t dropIndexes done" << endl;
+    LOG(1) << "\t dropIndexes done";
 
-    Top::get(txn->getClient()->getServiceContext()).collectionDropped(fullns);
+    Top::get(txn->getClient()->getServiceContext()).collectionDropped(fullns.toString());
 
-    s = _dbEntry->dropCollection(txn, fullns);
+    // We want to destroy the Collection object before telling the StorageEngine to destroy the
+    // RecordStore.
+    _clearCollectionCache(txn, fullns.toString(), "collection dropped");
 
-    // we want to do this always
-    _clearCollectionCache(txn, fullns, "collection dropped");
+    s = _dbEntry->dropCollection(txn, fullns.toString());
 
     if (!s.isOK())
         return s;
@@ -412,9 +416,7 @@ Status Database::dropCollection(OperationContext* txn, StringData fullns) {
         }
     }
 
-    auto opObserver = getGlobalServiceContext()->getOpObserver();
-    if (opObserver)
-        opObserver->onDropCollection(txn, nss);
+    getGlobalServiceContext()->getOpObserver()->onDropCollection(txn, fullns);
 
     return Status::OK();
 }
@@ -522,14 +524,15 @@ Status Database::createView(OperationContext* txn,
         return Status(ErrorCodes::InvalidNamespace,
                       str::stream() << "invalid namespace name for a view: " + nss.toString());
 
-    return _views.createView(txn, nss, viewOnNss, options.pipeline);
+    return _views.createView(txn, nss, viewOnNss, BSONArray(options.pipeline), options.collation);
 }
 
 
 Collection* Database::createCollection(OperationContext* txn,
                                        StringData ns,
                                        const CollectionOptions& options,
-                                       bool createIdIndex) {
+                                       bool createIdIndex,
+                                       const BSONObj& idIndex) {
     invariant(txn->lockState()->isDbLockedForMode(name(), MODE_X));
     invariant(!options.isView());
 
@@ -537,22 +540,27 @@ Collection* Database::createCollection(OperationContext* txn,
     _checkCanCreateCollection(nss, options);
     audit::logCreateCollection(&cc(), ns);
 
-    txn->recoveryUnit()->registerChange(new AddCollectionChange(txn, this, ns));
-
     Status status = _dbEntry->createCollection(txn, ns, options, true /*allocateDefaultSpace*/);
     massertNoTraceStatusOK(status);
 
-
+    txn->recoveryUnit()->registerChange(new AddCollectionChange(txn, this, ns));
     Collection* collection = _getOrCreateCollectionInstance(txn, ns);
     invariant(collection);
     _collections[ns] = collection;
+
+    BSONObj fullIdIndexSpec;
 
     if (createIdIndex) {
         if (collection->requiresIdIndex()) {
             if (options.autoIndexId == CollectionOptions::YES ||
                 options.autoIndexId == CollectionOptions::DEFAULT) {
+                const auto featureCompatibilityVersion =
+                    serverGlobalParams.featureCompatibility.version.load();
                 IndexCatalog* ic = collection->getIndexCatalog();
-                uassertStatusOK(ic->createIndexOnEmptyCollection(txn, ic->getDefaultIdIndexSpec()));
+                fullIdIndexSpec = uassertStatusOK(ic->createIndexOnEmptyCollection(
+                    txn,
+                    !idIndex.isEmpty() ? idIndex
+                                       : ic->getDefaultIdIndexSpec(featureCompatibilityVersion)));
             }
         }
 
@@ -561,9 +569,8 @@ Collection* Database::createCollection(OperationContext* txn,
         }
     }
 
-    auto opObserver = getGlobalServiceContext()->getOpObserver();
-    if (opObserver)
-        opObserver->onCreateCollection(txn, nss, options);
+    getGlobalServiceContext()->getOpObserver()->onCreateCollection(
+        txn, nss, options, fullIdIndexSpec);
 
     return collection;
 }
@@ -582,7 +589,7 @@ void dropAllDatabasesExceptLocal(OperationContext* txn) {
 
     if (n.size() == 0)
         return;
-    log() << "dropAllDatabasesExceptLocal " << n.size() << endl;
+    log() << "dropAllDatabasesExceptLocal " << n.size();
 
     repl::getGlobalReplicationCoordinator()->dropAllSnapshots();
     for (vector<string>::iterator i = n.begin(); i != n.end(); i++) {
@@ -615,21 +622,25 @@ void Database::dropDatabase(OperationContext* txn, Database* db) {
 
     audit::logDropDatabase(txn->getClient(), name);
 
+    for (auto&& coll : *db) {
+        Top::get(txn->getClient()->getServiceContext()).collectionDropped(coll->ns().ns(), true);
+    }
+
     dbHolder().close(txn, name);
     db = NULL;  // d is now deleted
 
-    getGlobalServiceContext()->getGlobalStorageEngine()->dropDatabase(txn, name);
+    MONGO_WRITE_CONFLICT_RETRY_LOOP_BEGIN {
+        getGlobalServiceContext()->getGlobalStorageEngine()->dropDatabase(txn, name);
+    }
+    MONGO_WRITE_CONFLICT_RETRY_LOOP_END(txn, "dropDatabase", name);
 }
 
-/** { ..., capped: true, size: ..., max: ... }
- * @param createDefaultIndexes - if false, defers id (and other) index creation.
- * @return true if successful
-*/
 Status userCreateNS(OperationContext* txn,
                     Database* db,
                     StringData ns,
                     BSONObj options,
-                    bool createDefaultIndexes) {
+                    bool createDefaultIndexes,
+                    const BSONObj& idIndex) {
     invariant(db);
 
     LOG(1) << "create collection " << ns << ' ' << options;
@@ -643,7 +654,7 @@ Status userCreateNS(OperationContext* txn,
         return Status(ErrorCodes::NamespaceExists,
                       str::stream() << "a collection '" << ns.toString() << "' already exists");
 
-    if (db->getViewCatalog()->lookup(ns))
+    if (db->getViewCatalog()->lookup(txn, ns))
         return Status(ErrorCodes::NamespaceExists,
                       str::stream() << "a view '" << ns.toString() << "' already exists");
 
@@ -694,7 +705,7 @@ Status userCreateNS(OperationContext* txn,
     if (collectionOptions.isView()) {
         uassertStatusOK(db->createView(txn, ns, collectionOptions));
     } else {
-        invariant(db->createCollection(txn, ns, collectionOptions, createDefaultIndexes));
+        invariant(db->createCollection(txn, ns, collectionOptions, createDefaultIndexes, idIndex));
     }
 
     return Status::OK();
